Minimal script to read sensor and print MQTT message
Reads directly from the Raspberry Pi UART (Pi TX on pin 8, RX on pin 10).

One line added to the SPS30.run_query() method to return the contents of
the MQTT message to the caller, which seems like it should be the
intended behavior. Otherwise, the method does not make the data available
at all when the save_data and push_mqtt options are False.